+++ core/src/chesscreator/chess/movement/MovementDirection.java
@@ -1,0 +1,4 @@
+package chesscreator.chess.movement;
+
+public class MovementDirection {
+}