+++ core/src/chesscreator/chess/piece/PieceColor.java
@@ -1,0 +1,5 @@
+package chesscreator.chess.piece;
+
+public enum PieceColor {
+    WHITE, BLACK
+}