+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [v.1.0-alpha] - 2021/8/20:
+### Added:
+- Chessboard, ChessPiece, PieceType and PieceColor objects! The chessboard can also parse fen strings.
+- Coordinates! Conversion between display coordinates (example: E4) and raw coordinates (example for E4: [4,4])
+- ConsoleChessRenderer, a chess renderer that simply displays the chessboard in a console.
+- A few tests that test the Chessboard and the Coordinates.
+
 ## [v.0.0-alpha] - 2021/8/17:
 ### Added:
 - Changelog file.